allosteric/requirements.txt
# Core data science
numpy==2.1.3
pandas==2.3.3
scipy==1.15.3
scikit-learn==1.7.2

# ML training
xgboost==3.2.0
optuna==4.7.0

# Deep learning / ESM-2
torch==2.6.0+cu124
fair-esm==2.0.0

# Protein structure analysis
biopython==1.86
prody==2.6.1
freesasa==2.2.1
pydssp==0.9.1
networkx==3.4.2

# Dataset I/O
h5py==3.15.1
joblib==1.5.3
